requirements.txt
attrs==21.4.0
getch==1.0
importlab==0.7
libcst==0.4.5
mutagen==1.45.1
mypy-extensions==0.4.3
networkx==2.8.3
ninja==1.10.2.3
pydot==1.4.2
pyparsing==3.0.9
python-vlc==3.0.16120
pytype==2022.6.23
PyYAML==6.0
tabulate==0.8.10
toml==0.10.2
typing-extensions==4.2.0
typing-inspect==0.7.1
